Edit float hour

+++ GtzTft.h
@@ -175,7 +175,7 @@ unsigned long printThankYou()
   return micros() - start;
 }
 
-unsigned long printStudent(String studentName, byte presentHour, byte previousHour, byte oldestHour, String studentStatus, String operation, String operationMessage, String subCode)
+unsigned long printStudent(String studentName, float presentHour, float previousHour, float oldestHour, String studentStatus, String operation, String operationMessage, String subCode)
 {
   GtzTft.fillScreen(ILI9341_WHITE);
   GtzTft.setTextColor(ILI9341_WHITE);
